升级Quartz到2.1.3,由于Quartz2.0以上的版本与Spring3.1以下的版本不兼容,所以把Spring升级到3.1.1版本;并修改了一些Bug!

+++ src/main/java/com/ewcms/scheduling/generate/job/channel/EwcmsExecutionChannelJob.java
@@ -59,7 +59,7 @@ public class EwcmsExecutionChannelJob extends BaseEwcmsExecutionJob {
         EwcmsJobChannelFacable ewcmsJobChannelFac = getEwcmsJobChannelFac();
         JobDataMap jobDataMap = jobContext.getTrigger().getJobDataMap();
 
-        int jobId = jobDataMap.getInt(JOB_DATA_KEY_DETAILS_ID);
+        Long jobId = jobDataMap.getLong(JOB_DATA_KEY_DETAILS_ID);
         EwcmsJobChannel ewcmsJobChannel = ewcmsJobChannelFac.getScheduledJobChannel(jobId);
         return ewcmsJobChannel;
     }
