+++ src/main/java/com/ewcms/scheduling/generate/job/crawler/EwcmsJobCrawlerFacable.java
@@ -22,7 +22,7 @@ public interface EwcmsJobCrawlerFacable {
 	 * @param jobId 任务编号
 	 * @return 采集器定时任务对象
 	 */
-	public EwcmsJobCrawler getScheduledJobCrawler(Integer jobId);
+	public EwcmsJobCrawler getScheduledJobCrawler(Long jobId);
 	
 	/**
 	 * 新增或修改采集器定时任务
@@ -32,7 +32,7 @@ public interface EwcmsJobCrawlerFacable {
 	 * @return 采集器定时任务编号
 	 * @throws BaseException
 	 */
-	public Integer saveOrUpdateJobCrawler(Long gatherId, PageDisplayVO vo) throws BaseException;
+	public Long saveOrUpdateJobCrawler(Long gatherId, PageDisplayVO vo) throws BaseException;
 	
 	/**
 	 * 通过采集器编号查询定时任务对象
