+++ src/main/java/com/ewcms/scheduling/generate/job/crawler/service/EwcmsJobCrawlerServiceable.java
@@ -24,7 +24,7 @@ public interface EwcmsJobCrawlerServiceable {
 	 * @param vo 定时设置对象
 	 * @return 定时工作任务编号
 	 */
-	public Integer saveOrUpdateJobCrawler(Long gatherId, PageDisplayVO vo) throws BaseException;
+	public Long saveOrUpdateJobCrawler(Long gatherId, PageDisplayVO vo) throws BaseException;
 	
 	/**
 	 * 通过任务编号查询采集器定时任务对象
@@ -32,7 +32,7 @@ public interface EwcmsJobCrawlerServiceable {
 	 * @param jobId 任务编号
 	 * @return 采集器定时任务对象
 	 */
-	public EwcmsJobCrawler getScheduledJobCrawler(Integer jobId);
+	public EwcmsJobCrawler getScheduledJobCrawler(Long jobId);
 	
 	/**
 	 * 通过采集器编号查询定时任务对象
